requirement.txt
Django==4.2.1
firebase-admin==6.6.0
requests==2.31.0
numpy==1.26.0
pandas==2.1.0
scikit-learn==1.3.0
matplotlib==3.8.0
seaborn==0.13.0
python-dotenv==1.0.0
setuptools==68.0.0
python-decouple==3.8
google-generativeai==0.8.3
langchain
langchain-core
langchain-google-firestore
langchain-google-genai
faiss-cpu
django-cors-headers==4.3.0
Pillow==10.0.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
